requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
apscheduler==3.10.4
vnstock3
pyyaml==6.0.1
python-dotenv==1.0.1
pytest==8.2.0
pytest-mock==3.14.0
